+++ changelog.md
@@ -8,6 +8,7 @@
 * Added a new Damage Type Tag `iguanatweaksexpanded:enchantments/protection/magic` defining which damage types are reduced by Magic Protection enchantment
 * Items can now be repaired again with nuggets
 * Passable foliage's Leaf Walker curse is now an enchantment
+* Farmer's delight integration (Season plant growth)
 * Fixed dupe glitch with soulbound
 * Fixed enchanted books from spawners having fake Curses